VMAF Integer Motion : Review#1 changes
Data bffer added in VMAFPicture to hold Offset subtracted YUV values.
will be used in motion,adm,and vif moduules.
Integer picture copy is removed.
Moved all functions related to integer convolution into integer_motion_functions.c file. integer_motion file is kept clean with only high level APIs (init,extarct, flush,close).
