+++ libvmaf/src/feature/common/convolution_internal.h
@@ -50,35 +50,6 @@ FORCE_INLINE inline float convolution_edge_s(bool horizontal, const float *filte
 	}
 	return accum;
 }
-/**
- * Works similar to floating-point function convolution_edge_s
-*/
-FORCE_INLINE inline int32_t integer_convolution_edge_s(bool horizontal, const uint16_t *filter, int filter_width, const int16_t *src, int width, int height, int stride, int i, int j)
-{
-	int radius = filter_width / 2;
-
-	int32_t accum = 0;
-	for (int k = 0; k < filter_width; ++k) {
-		int i_tap = horizontal ? i : i - radius + k;
-		int j_tap = horizontal ? j - radius + k : j;
-
-		// Handle edges by mirroring.
-		if (horizontal) {
-			if (j_tap < 0)
-				j_tap = -j_tap;
-			else if (j_tap >= width)
-				j_tap = width - (j_tap - width + 1);
-		} else {
-			if (i_tap < 0)
-				i_tap = -i_tap;
-			else if (i_tap >= height)
-				i_tap = height - (i_tap - height + 1);
-		}
-
-		accum += filter[k] * src[i_tap * stride + j_tap];
-	}
-	return accum;
-}
 
 FORCE_INLINE inline float convolution_edge_sq_s(bool horizontal, const float *filter, int filter_width, const float *src, int width, int height, int stride, int i, int j)
 {
